# Add WASM support