Closes #34!

This was a bit of a rabbit hole to get into, even though the changes are (thankfully) tiny but this adds WASM support to the crate. My use case was using a downstream crate in Cloudflare workers and this crate was causing panics because of its time APIs.

The PR uses [`web-time`](https://crates.io/crates/web-time) which was made specifically for this purpose; [`MonotonicClock`](https://docs.rs/governor/latest/governor/clock/struct.MonotonicClock.html) is now compatible with WASM environments at, *from what I understand*, no extra cost to non-WASM environments. 

> At the same time the library will simply re-export [std::time](https://doc.rust-lang.org/stable/std/time/) when not using the wasm32-unknown-unknown target and will not pull in any dependencies. 

^ From web-time's README

One small point that might cause confusion is when trying to type out a `RateLimiter` type which can look something like
`limiter: Arc<RateLimiter<NotKeyed, InMemoryState, MonotonicClock, NoOpMiddleware<web_time::Instant>>>`. If someone accidentally uses `std::time::Instant`, the errors are about Instant not implementing the traits that are outlined [here](https://docs.rs/governor/latest/governor/clock/index.html) which can be confusing to those trying to use this. For this I think for one documenting this use case is a must but also re-exporting `web_time::Instant` should come in handy.

My reasoning for the above is that while the existing docs are good and self explanatory, only someone that would try implementing his own Clock would likely take a look there; since this is an existing, internal Clock provided by the crate, there's a good chance most people using it won't have read that part of the docs.

You will also notice this PR adds `getrandom = { version = "0.2", features = ["js"] }`, this is required to get `rand` to compile in a WASM environment (see [here](https://github.com/rust-random/rand?tab=readme-ov-file#wasm-support)).

Lastly, nothing of the included changes is put behind a feature flag but I personally don't think any of the changes could cause issues. Let me know what you think!