@antifuchs 2025-01-23
Wow-eeeee, this is amazing & exciting. I love that it's basically a 3-line diff hiding a _ton_ of hard work by many people, especially you (:

I'll try to read up on what the implications are, but I'm super excited to get this merged asap - thank you for this!