@antifuchs 2025-01-23
Looks like my upstream github workflows use some deprecated version or something, which will prevent some runs from succeeding. I'm looking into it!