@AntoniosBarotsis 2025-01-23
Looking at the performance from [this pr](https://github.com/boinkor-net/governor/actions/runs/12915207651/job/36066064341) vs [an older run from master](https://github.com/boinkor-net/governor/actions/runs/12259176790/job/34200710835) performance seems to indeed be the same which is great!

This PR:
```
Benchmarking realtime_clock/mostly_allow/MonotonicClock: Collecting 100 samples in estimated 5.0001 s (126M iterations)
Benchmarking realtime_clock/mostly_allow/MonotonicClock: Analyzing
realtime_clock/mostly_allow/MonotonicClock
                        time:   [39.484 ns 39.570 ns 39.712 ns]
                        thrpt:  [25.181 Melem/s 25.272 Melem/s 25.326 Melem/s]
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe
```

Master:

```
Benchmarking realtime_clock/mostly_allow/MonotonicClock: Collecting 100 samples in estimated 5.0000 s (129M iterations)
Benchmarking realtime_clock/mostly_allow/MonotonicClock: Analyzing
realtime_clock/mostly_allow/MonotonicClock
                        time:   [38.674 ns 38.693 ns 38.719 ns]
                        thrpt:  [25.827 Melem/s 25.844 Melem/s 25.857 Melem/s]
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  6 (6.00%) high severe
```

So it is an effectively "free" change for anyone using the crate same as before, outside of WASM.