@AntoniosBarotsis 2025-01-23
I'm on Windows and only 1 test is failing by being too slow but I'm pretty sure that's because I use WSL on a sluggish laptop, everything else passes so once the CI upload artifact v3 thing is resolved everything should be green!

```
failures:

---- proceeds_keyed_n stdout ----
thread 'proceeds_keyed_n' panicked at governor/tests/future.rs:99:5:
assertion failed: `(left <= right) but here (left > right)`
  left: `238.7µs`,
 right: `200µs`
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace


failures:
    proceeds_keyed_n
```