@antifuchs 2025-01-23
Looks like there are just no-std builds left due to a downstream dep not being fully no-std compatible (and a cargo-deny migration); I'll fix those on master in a bit.