@antifuchs 2025-01-23
Oh, just one thought: How did you try this (what toolchains / tools)? I'd love to add that configuration to CI so we keep it working.