@AntoniosBarotsis 2025-01-23
Testing is a bit weird because from what I understand your code might need to be evaluated for it to crash, i.e. just because something compiles *for WASM* does not mean it will not panic when targeting WASM.

What I did was follow [these instructions for the HTTP example](https://github.com/cloudflare/workers-rs?tab=readme-ov-file#http-feature) and just run `worker-build --release` followed by `npx wrangler dev` and visiting the URL.

To illustrate what I mentioned earlier about things compiling even though they might lead to a panic, the following code compiles just fine

```rs
use axum::{routing::get, Router};
use tower_service::Service;
use worker::*;

fn router() -> Router {
    Router::new().route("/", get(root))
}

#[event(fetch)]
async fn fetch(
    req: HttpRequest,
    _env: Env,
    _ctx: Context,
) -> Result<axum::http::Response<axum::body::Body>> {
    console_error_panic_hook::set_once();

    let tmp = std::time::Instant::now();
  
    Ok(router().call(req).await?)
}

pub async fn root() -> &'static str {
    "Hello Axum!"
}
```

But once you visit the URL it panics because of the use of `Instant`. There has to be better ways of testing for stuff like this (my eye caught [this](https://rustwasm.github.io/wasm-bindgen/wasm-bindgen-test/index.html) for example which has to be what you're supposed to do) but since I haven't done anything WASM related before, I'm not too sure.