@AntoniosBarotsis 2025-01-23
I tried adding some tests via `wasm-bindgen-test` but it seems rather annoying getting them to actually work. At first they crash because of `proptest`, remove that and then `rayon` causes issues. I'm not sure if there's a way to *only* compile what's needed for those tests and nothing else, otherwise you'd probably need a separate crate for testing 🤔