@antifuchs 2025-01-24
Ah, yeah, I suppose we can make a sub-crate here that is only for testing. Thankfully the repo is already organized as a workspace.

That's just a follow-up item, though. No need to get busy on that in this PR! (: