@AntoniosBarotsis 2025-01-24
If anything interesting pops up in the testing phase let me know as I'll also need to write some wasm tests in my own project 😅 thanks!